refactor: inline `runOpen` util
